backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy>=2.0.30
# psycopg2-binary==2.9.6 (PostgreSQL driver - disabled for SQLite)
python-jose==3.3.0
firebase-admin==6.0.0
redis==5.0.0
celery==5.3.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.38.0
reportlab>=4.0.0
boto3>=1.34.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
googlemaps>=4.10.0